% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_column}
\alias{create_column}
\title{Create column}
\usage{
create_column(clmnList, frame)
}
\arguments{
\item{clmnList}{A list detailing the schema for the column}

\item{frame}{The dataframe where the column will be inserted into}
}
\value{
A Dataframe
}
\description{
This function creates a column in the 'frame' dataframe based on
 the specification provided in the 'clmnList' list.
}
\author{
Nathanael Rolim \email{Nathanael.Rolim@gamail.com}
}
