% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_tables}
\alias{create_tables}
\title{Create Tables}
\usage{
create_tables(list)
}
\arguments{
\item{list}{The list object created by the YAML file.}
}
\value{
List
}
\description{
A function that creates the tables provided in the configuration schema. Normally,
 it returns dataframes.
}
