% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{import_schema}
\alias{import_schema}
\title{Import schema}
\usage{
import_schema(file = "")
}
\arguments{
\item{file}{The YAML schema file}
}
\description{
This function reads and parses the YAML schema file. It also verifies
 if the schema is correct and if all arguments are valid.
}
\examples{
 \dontrun{
 importParse("_faker.yaml")
 }

}
\author{
Nathanael Rolim \email{Nathanael.Rolim@gamail.com}
}
