% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First.R
\name{importParse}
\alias{importParse}
\title{Import and Parse YAML file}
\usage{
importParse(file = "")
}
\description{
This function reads the YAML file.
}
\author{
Nathanael Rolim \email{Nathanael.Rolim@gamail.com}
}
