% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{verify_schema}
\alias{verify_schema}
\title{Verify Schema}
\usage{
verify_schema(list)
}
\arguments{
\item{list}{The list object created by the YAML file.}
}
\value{
Returns `TRUE` if the schema parses correctly, or fails otherwise.
}
\description{
This function takes a list generated by importing a YAML file and
 verifies if it matches the necessary schema.
}
\details{
A function the verify whether a given schema is correct for use with
 the `data.fakeR` package. It performs the following checks:
 \itemize{
  \item 1 - First, it verifies whether a list named `tables` exists. If it does, then it verifies whether this list contains at least one element.
  \item 2 - Second, it checks if each table: (a) is named, (b) has a given number of rows, and (c) has at least one column.
  \item 3 - Finally, it loops over all columns to determine if they: (a) have a name, (b) have a valid data type (given the column type), and (c) have a valid column type.
 }
}
\examples{
 \dontrun{
  schm <- yaml::read_yaml("schema.yaml")
  verify_schema(schm)
 }

}
\author{
Nathanael Rolim \email{Nathanael.Rolim@gamail.com}
}
