% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{create_tables}
\alias{create_tables}
\title{Create Tables}
\usage{
create_tables(list)
}
\arguments{
\item{list}{The list object created by the YAML file.}
}
\value{
A list of dataframes specified in the schema list.
}
\description{
A function that creates the tables provided in the configuration schema. Normally,
 it returns dataframes. This function can be used on its own, but it is better if it is called
 after `import_schema`.s
}
