% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{import_schema}
\alias{import_schema}
\title{Import schema}
\usage{
import_schema(file = "")
}
\arguments{
\item{file}{The YAML schema file}
}
\value{
A list containing the tables schema.
}
\description{
A wrapper function for reading a YAML schema file. It calls `verify_schema` to test whether the
 imported schema is correct for use with this package.
}
\examples{
 \dontrun{
 importParse("_faker.yaml")
 }

}
\author{
Nathanael Rolim \email{Nathanael.Rolim@gamail.com}
}
