% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{verify_schema}
\alias{verify_schema}
\title{Verify Schema}
\usage{
verify_schema(list)
}
\arguments{
\item{list}{The list object created by the YAML file.}
}
\value{
Logical
}
\description{
This function takes a list generated by importing a YAML file and
 verifies if it matches the necessary schema.
}
\examples{
 \dontrun{
  schm <- yaml::read_yaml("schema.yaml")
  verify_schema(schm)
 }

}
\author{
Nathanael Rolim \email{Nathanael.Rolim@gamail.com}
}
